build: fix global filter story, babel cleanup
