+++ .babelrc.json
@@ -1,19 +1,5 @@
 {
   "sourceType": "unambiguous",
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "chrome": 100
-        }
-      }
-    ],
-    "@babel/preset-typescript"
-  ],
-  "plugins": [
-    ["@babel/plugin-proposal-decorators", { "version": "2023-11" }],
-    "@babel/plugin-transform-class-properties",
-    "@babel/plugin-transform-class-static-block"
-  ]
+  "presets": ["@babel/preset-env", "@babel/preset-typescript"],
+  "plugins": [["@babel/plugin-proposal-decorators", { "version": "2023-11" }]]
 }
